requirements.txt
altgraph==0.17.3
antiorm==1.2.1
anyio==3.6.2
auto-py-to-exe==2.26.1
bcrypt==4.0.1
bottle==0.12.23
bottle-websocket==0.2.9
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
cryptography==38.0.4
db==0.1.1
distlib==0.3.6
ecdsa==0.18.0
Eel==0.14.0
fastapi==0.87.0
filelock==3.8.2
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
h11==0.14.0
identify==2.5.9
idna==3.4
nodeenv==1.7.0
passlib==1.7.4
platformdirs==2.6.0
pre-commit==2.20.0
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.0.9
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
toml==0.10.2
typing-extensions==4.4.0
uvicorn==0.20.0
virtualenv==20.17.1
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2
